updated contacts page

+++ frontend/src/Components/Contacts.jsx
@@ -1,9 +1,12 @@
 import { useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
+
 import logo from "../assets/logo.png";
 import deleteImg from "../assets/delete.png";
+import malePic from "../assets/male.png";
+import femalePic from "../assets/female.png";
 import { useLogout } from "../Hooks/useLogout";
 import logoutImg from "../assets/logout.png";
-import { Link } from "react-router-dom";
 import { useAuthContext } from "../Hooks/useAuthContext";
 
 function Contacts({ data }) {
@@ -15,17 +18,24 @@ function Contacts({ data }) {
     navigate("/new-contact");
   }
 
-  async function HandleDeleteBtnClick(_id) {
-    const res = await fetch("http://localhost:4000/" + _id, {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${user.user.token}`,
-      },
-    });
-    const json = await res.json();
+  async function HandleDeleteBtnClick(_id, name) {
+    const confirmDelete = window.confirm(
+      `Do you really want to delete ${name}`
+    );
 
-    if (json) {
-      window.location.reload();
+    if (confirmDelete) {
+      const res = await fetch("http://localhost:4000/" + _id, {
+        method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.user.token}`,
+        },
+      });
+      const json = await res.json();
+
+      if (json) {
+        window.location.reload();
+        alert(`${name} has been deleted.`);
+      }
     }
   }
 
@@ -49,10 +59,13 @@ function Contacts({ data }) {
         </button>
       </div>
 
-      <div className="ml-52 mr-52 mt-4 rounded-2xl  bg-white overflow-auto max-h-[350px]">
+      <div className="ml-52 mr-52 mt-4 rounded-2xl bg-white overflow-auto max-h-[350px] overflow-x-auto">
         <table className="w-full text-base text-left rtl:text-right text-twc-green">
           <thead className="text-xs text-twc-green uppercase bg-white ">
             <tr>
+              <th scope="col" className="px-6 py-3">
+                profile pic
+              </th>
               <th scope="col" className="px-6 py-3">
                 full name
               </th>
@@ -65,26 +78,40 @@ function Contacts({ data }) {
               <th scope="col" className="px-6 py-3">
                 phone number
               </th>
+              <th scope="col" className="px-4 py-3">
+                delete
+              </th>
             </tr>
           </thead>
           <tbody>
             {data.map((contact, index) => (
               <tr key={index} className="bg-white ">
+                <td className="px-6 py-4">
+                  {contact.gender !== "male" ? (
+                    <img src={femalePic} alt="female profile pic" />
+                  ) : (
+                    <img src={malePic} alt="male profile pic" />
+                  )}
+                </td>
                 <td className="px-6 py-4">{contact.name}</td>
                 <td className="px-6 py-4">{contact.gender}</td>
                 <td className="px-6 py-4">{contact.email}</td>
                 <td className="px-6 py-4">{contact.phoneNumber}</td>
-                <button
-                  className="mt-4"
-                  onClick={() => HandleDeleteBtnClick(contact._id)}
-                >
-                  <img src={deleteImg} alt="delete image" />
-                </button>
+                <td className="px-6 py-4">
+                  <button
+                    onClick={() =>
+                      HandleDeleteBtnClick(contact._id, contact.name)
+                    }
+                  >
+                    <img src={deleteImg} alt="delete image" />
+                  </button>
+                </td>
               </tr>
             ))}
           </tbody>
         </table>
       </div>
+
       <div
         className="fixed bottom-10 right-10 flex items-center ml-[500px] space-x-2"
         onClick={handleClick}
